created usable module for abr log reading
