+++ hardware-testing/hardware_testing/abr_tools/abr_read_logs.py
@@ -1,14 +1,11 @@
 """Read ABR run logs and save data to ABR testing csv and google sheet."""
-from .abr_run_logs import get_run_ids_from_storage, get_unseen_run_ids
-from .error_levels import ERROR_LEVELS_PATH
-from typing import Set, Dict, Tuple, Any, List
+from typing import Set, Dict, Any
 import argparse
 import os
-import csv
 import json
 import sys
 from datetime import datetime, timedelta
-import time as t
+from hardware_testing.abr_tools.read_robot_logs import ReadRobotLogs
 
 
 def get_modules(file_results: Dict[str, str]) -> Dict[str, Any]:
@@ -28,59 +25,6 @@ def get_modules(file_results: Dict[str, str]) -> Dict[str, Any]:
                 all_modules[module["model"]] = "EMPTYSN"
 
     return all_modules
-
-
-def get_error_info(file_results: Dict[str, Any]) -> Tuple[int, str, str, str, str]:
-    """Determines if errors exist in run log and documents them."""
-    error_levels = []
-    # Read error levels file
-    with open(ERROR_LEVELS_PATH, "r") as error_file:
-        error_levels = list(csv.reader(error_file))
-    num_of_errors = len(file_results["errors"])
-    if num_of_errors == 0:
-        error_type = ""
-        error_code = ""
-        error_instrument = ""
-        error_level = ""
-        return 0, error_type, error_code, error_instrument, error_level
-    commands_of_run: List[Dict[str, Any]] = file_results.get("commands", [])
-    run_command_error: Dict[str, Any] = commands_of_run[-1]
-    error_str: int = len(run_command_error.get("error", ""))
-    if error_str > 1:
-        error_type = run_command_error["error"].get("errorType", "")
-        error_code = run_command_error["error"].get("errorCode", "")
-        try:
-            # Instrument Error
-            error_instrument = run_command_error["error"]["errorInfo"]["node"]
-        except KeyError:
-            # Module Error
-            error_instrument = run_command_error["error"]["errorInfo"].get("port", "")
-    else:
-        error_type = file_results["errors"][0]["errorType"]
-        print(error_type)
-        error_code = file_results["errors"][0]["errorCode"]
-        error_instrument = file_results["errors"][0]["detail"]
-    for error in error_levels:
-        code_error = error[1]
-        if code_error == error_code:
-            error_level = error[4]
-
-    return num_of_errors, error_type, error_code, error_instrument, error_level
-
-
-def create_abr_data_sheet(storage_directory: str, file_name: str, headers: List) -> str:
-    """Creates csv file to log ABR data."""
-    file_name_csv = file_name + ".csv"
-    print(file_name_csv)
-    sheet_location = os.path.join(storage_directory, file_name_csv)
-    if os.path.exists(sheet_location):
-        print(f"File {sheet_location} located. Not overwriting.")
-    else:
-        with open(sheet_location, "w") as csvfile:
-            writer = csv.DictWriter(csvfile, fieldnames=headers)
-            writer.writeheader()
-            print(f"Created file. Located: {sheet_location}.")
-    return file_name_csv
 
 
 def create_data_dictionary(
@@ -109,7 +53,7 @@ def create_data_dictionary(
                 error_code,
                 error_instrument,
                 error_level,
-            ) = get_error_info(file_results)
+            ) = ReadRobotLogs.get_error_info(file_results)
             all_modules = get_modules(file_results)
 
             start_time_str, complete_time_str, start_date, run_time_min = (
@@ -160,52 +104,6 @@ def create_data_dictionary(
                 os.remove(file_path)
                 print(f"Run ID: {run_id} has a run time of 0 minutes. Run removed.")
     return runs_and_robots
-
-
-def read_abr_data_sheet(
-    storage_directory: str, file_name_csv: str, google_sheet: Any
-) -> Set[str]:
-    """Reads current run sheet to determine what new run data should be added."""
-    print(file_name_csv)
-    sheet_location = os.path.join(storage_directory, file_name_csv)
-    runs_in_sheet = set()
-    # Read the CSV file
-    with open(sheet_location, "r") as csv_start:
-        data = csv.DictReader(csv_start)
-        headers = data.fieldnames
-        if headers is not None:
-            for row in data:
-                run_id = row[headers[1]]
-                runs_in_sheet.add(run_id)
-        print(f"There are {str(len(runs_in_sheet))} runs documented in the ABR sheet.")
-    # Read Google Sheet
-    if google_sheet.creditals.access_token_expired:
-        google_sheet.gc.login()
-    google_sheet.write_header(headers)
-    google_sheet.update_row_index()
-    return runs_in_sheet
-
-
-def write_to_abr_sheet(
-    runs_and_robots: Dict[Any, Dict[str, Any]],
-    storage_directory: str,
-    file_name_csv: str,
-    google_sheet: Any,
-) -> None:
-    """Write dict of data to abr csv."""
-    sheet_location = os.path.join(storage_directory, file_name_csv)
-    list_of_runs = list(runs_and_robots.keys())
-    with open(sheet_location, "a", newline="") as f:
-        writer = csv.writer(f)
-        for run in range(len(list_of_runs)):
-            row = runs_and_robots[list_of_runs[run]].values()
-            row_list = list(row)
-            writer.writerow(row_list)
-            if google_sheet.creditals.access_token_expired:
-                google_sheet.gc.login()
-            google_sheet.update_row_index()
-            google_sheet.write_to_row(row_list)
-            t.sleep(3)
 
 
 if __name__ == "__main__":
@@ -273,9 +171,15 @@ if __name__ == "__main__":
         "magneticBlockV1",
         "thermocyclerModuleV2",
     ]
-    runs_from_storage = get_run_ids_from_storage(storage_directory)
-    file_name_csv = create_abr_data_sheet(storage_directory, file_name, headers)
-    runs_in_sheet = read_abr_data_sheet(storage_directory, file_name_csv, google_sheet)
-    runs_to_save = get_unseen_run_ids(runs_from_storage, runs_in_sheet)
+    runs_from_storage = ReadRobotLogs.get_run_ids_from_storage(storage_directory)
+    file_name_csv = ReadRobotLogs.create_abr_data_sheet(
+        storage_directory, file_name, headers
+    )
+    runs_in_sheet = ReadRobotLogs.read_abr_data_sheet(
+        storage_directory, file_name_csv, google_sheet
+    )
+    runs_to_save = ReadRobotLogs.get_unseen_run_ids(runs_from_storage, runs_in_sheet)
     runs_and_robots = create_data_dictionary(runs_to_save, storage_directory)
-    write_to_abr_sheet(runs_and_robots, storage_directory, file_name_csv, google_sheet)
+    ReadRobotLogs.write_to_abr_sheet(
+        runs_and_robots, storage_directory, file_name_csv, google_sheet
+    )
